finish author modul

+++ public/js/admin/news/author/index.js
@@ -1,42 +1,59 @@
-$(document).on('click', '.delete__link', function() {
+$(document).on('click', '.delete__link', function () {
     const author_id = $(this).data('id');
     const row = $(this).closest('tr');
     const rows = $('.table__body tr');
 
-    $.ajax({
-        type: 'DELETE',
-        url: 'author/' + author_id,
-        headers: {
-            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-        },
-        success: function(response) {
+    // Use SweetAlert for confirmation
+    Swal.fire({
+        title: delete_author,
+        text: confirm_author_delete,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: 'var(--green-btn)',
+        cancelButtonColor: 'var(--delete)',
+        confirmButtonText: 'დიახ',
+        cancelButtonText: 'არა'
+    }).then((result) => {
+        if (result.isConfirmed) {
 
-            row.fadeOut('slow', function() {
-                $(this).remove();
-                rows.removeClass('blur');
-            });
+            $.ajax({
+                type: 'DELETE',
+                url: 'author/' + author_id,
+                headers: {
+                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                },
+                success: function (response) {
+                    row.fadeOut('slow', function () {
+                        $(this).remove();
+                        rows.removeClass('blur');
+                    });
 
-            const Toast = Swal.mixin({
-                toast: true,
-                position: 'top-end',
-                showConfirmButton: false,
-                timer: 1000,
-                timerProgressBar: true,
-                didOpen: (toast) => {
-                    toast.addEventListener('mouseenter', Swal.stopTimer)
-                    toast.addEventListener('mouseleave', Swal.resumeTimer)
+                    const Toast = Swal.mixin({
+                        toast: true,
+                        position: 'top-end',
+                        showConfirmButton: false,
+                        timer: 1000,
+                        timerProgressBar: true,
+                        didOpen: (toast) => {
+                            toast.addEventListener('mouseenter', Swal.stopTimer)
+                            toast.addEventListener('mouseleave', Swal.resumeTimer)
+                        }
+                    })
+
+                    Toast.fire({
+                        icon: 'success',
+                        title: 'ავტორი წარმატებით წაიშალა'
+                    });
+
+                    setTimeout(function (){
+                        location.reload()
+                    }, 1000);
+
+                },
+                error: function (response) {
+                    $('#delete_message').html(response.responseJSON.error);
                 }
-            })
-
-            Toast.fire({
-                icon: 'success',
-                title: 'ავტორი წარმატებით წაიშალა'
-            })
-
-        },
-        error: function(response) {
-            $('#delete_message').html(response.responseJSON.error);
+            });
         }
     });
 });
-
